:construction: Move extension method to a separate class
