+++ Toggl.Daneel/Toggl.Daneel.csproj
@@ -1178,6 +1178,7 @@
     <Compile Include="Views\EntityCreation\CreateEntityViewCell.designer.cs">
       <DependentUpon>CreateEntityViewCell.cs</DependentUpon>
     </Compile>
+    <Compile Include="Binding\AnimatedTableViewReloadingBinder.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Toggl.Foundation.MvvmCross\Toggl.Foundation.MvvmCross.csproj">
